requirements.txt
numpy==1.19.5
opencv_python==4.5.5.62
Pillow==9.0.0
PyAutoGUI==0.9.53
